Move `FeedImageDataStoreSpy` to a cache shared scope
